Migrate from CircleCI to GitHub Actions for testing and releases, enable linting and changelog (#866)
Reference: https://github.com/hashicorp/terraform-plugin-sdk/issues/794

All testing and linting workflows verified as successful via:

```shell
act -s GITHUB_TOKEN=$GITHUB_TOKEN pull_request
```

+++ helper/schema/field_writer_map.go
@@ -271,7 +271,7 @@ func (w *MapFieldWriter) setPrimitive(
 		if err := mapstructure.Decode(v, &n); err != nil {
 			return fmt.Errorf("%s: %s", k, err)
 		}
-		set = strconv.FormatFloat(float64(n), 'G', -1, 64)
+		set = strconv.FormatFloat(n, 'G', -1, 64)
 	default:
 		return fmt.Errorf("Unknown type: %#v", schema.Type)
 	}
